Revert "Code climate issue"
