--- project-cli/files/collection.txt
@@ -1,3 +1,0 @@
-This projects includes files which is specified at the below
-
-Happy Coding!
