+++ project-cli/helper_scripts/templates/components/class.mustache
