+++ project-cli/helper_scripts/templates/components/functional.mustache
