+++ project-cli/helper_scripts/templates/components/index.mustache
