+++ project-cli/helper_scripts/templates/interfaces/component-index.d.mustache
