+++ project-cli/helper_scripts/templates/interfaces/component.d.mustache
