+++ project-cli/helper_scripts/templates/interfaces/page-index.d.mustache
