+++ project-cli/helper_scripts/templates/interfaces/redux-import.d.mustache
