+++ project-cli/helper_scripts/templates/interfaces/redux-store.d.mustache
