+++ project-cli/helper_scripts/templates/reducers/action-index.mustache
