+++ project-cli/helper_scripts/templates/reducers/action.mustache
