+++ project-cli/helper_scripts/templates/reducers/index.mustache
