+++ project-cli/helper_scripts/templates/reducers/reducer.mustache
