+++ project-cli/helper_scripts/templates/reducers/store.mustache
