+++ project-cli/helper_scripts/templates/routes.mustache
