+++ project-cli/helper_scripts/templates/styles.mustache
