--- project-cli/src/definations/functional-component.ts
@@ -1,20 +1,0 @@
-import * as inquirer from 'inquirer';
-import { DefinationsModel } from './Defination';
-import { Helper } from './helper';
-import { Common } from './common';
-
-export const funcComp = {
-	showQuestions: async (): Promise<void> => {
-		const answers: DefinationsModel.IAnswers =
-			await inquirer.prompt<{ fileName: string, isHaveStyle: boolean }>(Common.funcCompQuestions);
-
-		answers.fileName = answers.fileName.replace(/\b\w/g, foo => foo.toUpperCase());
-		answers.lowerFileName = answers.fileName.replace(/\b\w/g, foo => foo.toLowerCase());
-
-		Helper.createFuncComponent(answers);
-
-		if (answers.isHaveStyle) {
-			Helper.createStyle(answers);
-		}
-	}
-};
