--- project-cli/test/unit/helper.spec.ts
@@ -1,13 +1,0 @@
-
-
-import { Config } from '../../config';
-import { DefinationsModel } from '../../src/definations/Defination';
-import { Helper } from '../../src/definations/helper';
-import * as path from 'path';
-
-describe('Test Helper constructor', () => {
-
-	 test('', () => {
-	});
-
-});
